requirements.txt
fastmcp>=0.2.0
httpx>=0.27.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
